+++ Genelife.Domain/Events/Company/EmployeeProductivityUpdated.cs
@@ -1,3 +1,3 @@
 namespace Genelife.Domain.Events.Company;
 
-public record EmployeeProductivityUpdated(Guid CompanyId, Guid HumanId, decimal ProductivityScore);
+public record EmployeeProductivityUpdated(Guid CompanyId, Guid HumanId, float ProductivityScore);
