+++ Genelife.Domain/Events/Company/SalaryPaid.cs
@@ -2,4 +2,4 @@ using MassTransit;
 
 namespace Genelife.Domain.Events.Company;
 
-public record SalaryPaid(Guid CorrelationId, decimal Amount, decimal TaxDeducted) : CorrelatedBy<Guid>;
+public record SalaryPaid(Guid CorrelationId, float Amount, float TaxDeducted) : CorrelatedBy<Guid>;
